backend/requirements.txt
Django==4.2.7
djangorestframework==3.14.0
django-cors-headers==4.3.1
psycopg2-binary==2.9.11
cryptography==41.0.7
python-dotenv==1.0.0
PyJWT==2.8.0
redis==5.0.1
celery==5.3.4
django-celery-beat==2.5.0
django-celery-results==2.5.1
flower==2.0.1
django-redis==5.4.0
django-filter==23.5
